# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMDB API",
        "version": "3.0"
    },
    "servers": [
        {
            "url": "/api/v2",
            "description": "API v2"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Authenticate Existing User",
                "operationId": "auth/login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Username",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User profile data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ResponceItem"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                },
                                                "extdata": {
                                                    "description": "Authorization TOKEN",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/auth/refresh": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Refrsh Existing User Token",
                "operationId": "auth/refresh",
                "responses": {
                    "200": {
                        "description": "User profile data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ResponceItem"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                },
                                                "extdata": {
                                                    "description": "New Authorization TOKEN",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTToken": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Logout authorized user",
                "operationId": "auth/logout",
                "responses": {
                    "200": {
                        "description": "Logout complete"
                    }
                },
                "security": [
                    {
                        "JWTToken": []
                    }
                ]
            }
        },
        "/auth/profile": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Return User profile data",
                "operationId": "auth/profile",
                "responses": {
                    "200": {
                        "description": "User profile data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ResponceItem"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTToken": []
                    }
                ]
            }
        },
        "/auth/access": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Return User access list",
                "operationId": "auth/access",
                "responses": {
                    "200": {
                        "description": "User access list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ResponceItem"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Access"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ResponceItem": {
                "allOf": [
                    {
                        "properties": {
                            "status": {
                                "description": "Status code",
                                "type": "integer",
                                "default": 200
                            },
                            "message": {
                                "description": "Information message",
                                "type": "string"
                            },
                            "data": {
                                "description": "Data",
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            },
                            "extdata": {
                                "description": "ExtData",
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            },
                            "debug": {
                                "description": "Debug information",
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "User": {
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "username": {
                        "description": "Username",
                        "type": "string"
                    },
                    "fio": {
                        "description": "Last First Subling (names)",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string"
                    },
                    "admin": {
                        "description": "isAdmin",
                        "type": "integer"
                    },
                    "flag": {
                        "description": "isActive",
                        "type": "integer"
                    },
                    "ldap": {
                        "description": "isLDAP",
                        "type": "integer"
                    },
                    "date_logon": {
                        "description": "Date of last logon",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Access": {
                "properties": {
                    "admin": {
                        "description": "is admin",
                        "type": "boolean"
                    },
                    "access": {
                        "description": "list of access",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "groups": {
                        "description": "list of groups",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "JWTToken": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "JWTToken": []
        }
    ],
...
}